Alter interface to reduce confusion (#5)
This PR tweaks the fn copy interface to take an item: &T as an input and return a T as an output. The returned value should not be dropped (unsafe to do so). This is meant to make the interface a bit more clear, in that there are fewer *mut T things flopping around.

There is the risk that it is "more unsafe" because it is easy to call copy(item) and just walk away; arguably the result should be wrapped in ManuallyDrop but the method is still unsafe and there is just a bit more screwing around with the result. There is also the risk that performance could now be worse, because we need to rely on Rust and LLVM to recover memcpy performance.