Only indicate allocated data in heap_size
Prior to this change, calling heap_size on a region would reveal
information about allocations that were not filled, i.e., did not own a
valid heap allocation. This was not a problem as long as the caller
did not count the number of callback invocations, but would invoke the
callback more times than there are owned allocations.

After this change, we only call the callback if the allocation exists,
for example if a vector has a capacity. This is aligned with the
specification.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>